Consolidate testing
