+++ python/cuda_cccl/tests/compute/test_binary_search.py
@@ -6,7 +6,13 @@ import numpy as np
 import pytest
 
 import cuda.compute
-from cuda.compute import OpKind
+from cuda.compute import (
+    OpKind,
+    make_lower_bound,
+    make_upper_bound,
+    serialize,
+)
+from cuda.compute.algorithms import load_lower_bound, load_upper_bound
 
 DTYPE_LIST = [
     np.int32,
@@ -223,3 +229,55 @@ def test_binary_search_requires_pointer_sized_output():
             num_values=len(d_values),
             d_out=d_out,
         )
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_lower_bound_round_trip():
+    h_data = np.array([1, 3, 3, 5, 7, 9], dtype=np.int32)
+    h_values = np.array([0, 3, 4, 10], dtype=np.int32)
+    d_data = cp.asarray(h_data)
+    d_values = cp.asarray(h_values)
+    d_out = cp.empty(len(h_values), dtype=np.uintp)
+
+    builder = make_lower_bound(d_data=d_data, d_values=d_values, d_out=d_out)
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = load_lower_bound(blob)
+    loaded(
+        d_data=d_data,
+        num_items=len(d_data),
+        d_values=d_values,
+        num_values=len(d_values),
+        d_out=d_out,
+        comp=None,
+    )
+
+    expected = np.searchsorted(h_data, h_values, side="left").astype(np.uintp)
+    np.testing.assert_array_equal(d_out.get(), expected)
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_upper_bound_round_trip():
+    h_data = np.array([1, 3, 3, 5, 7, 9], dtype=np.int32)
+    h_values = np.array([0, 3, 4, 10], dtype=np.int32)
+    d_data = cp.asarray(h_data)
+    d_values = cp.asarray(h_values)
+    d_out = cp.empty(len(h_values), dtype=np.uintp)
+
+    builder = make_upper_bound(d_data=d_data, d_values=d_values, d_out=d_out)
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = load_upper_bound(blob)
+    loaded(
+        d_data=d_data,
+        num_items=len(d_data),
+        d_values=d_values,
+        num_values=len(d_values),
+        d_out=d_out,
+        comp=None,
+    )
+
+    expected = np.searchsorted(h_data, h_values, side="right").astype(np.uintp)
+    np.testing.assert_array_equal(d_out.get(), expected)
