+++ python/cuda_cccl/tests/compute/test_histogram.py
@@ -12,7 +12,11 @@ import cuda.compute
 from cuda.compute import (
     ConstantIterator,
     CountingIterator,
+    deserialize,
+    make_histogram_even,
+    serialize,
 )
+from cuda.compute._utils.temp_storage_buffer import TempStorageBuffer
 
 DTYPE_LIST = [
     np.uint8,
@@ -548,3 +552,77 @@ def test_make_histogram_even_rejects_mismatched_bound_dtypes():
             h_upper_level=np.array([8], dtype=np.float32),
             num_samples=num_samples,
         )
+
+
+def _run(
+    histogram,
+    *,
+    d_samples,
+    d_histogram,
+    h_num_output_levels,
+    h_lower_level,
+    h_upper_level,
+    num_samples,
+):
+    bytes_needed = histogram(
+        temp_storage=None,
+        d_samples=d_samples,
+        d_histogram=d_histogram,
+        h_num_output_levels=h_num_output_levels,
+        h_lower_level=h_lower_level,
+        h_upper_level=h_upper_level,
+        num_samples=num_samples,
+    )
+    tmp = TempStorageBuffer(bytes_needed, None)
+    histogram(
+        temp_storage=tmp,
+        d_samples=d_samples,
+        d_histogram=d_histogram,
+        h_num_output_levels=h_num_output_levels,
+        h_lower_level=h_lower_level,
+        h_upper_level=h_upper_level,
+        num_samples=num_samples,
+    )
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_histogram_even_round_trip():
+    num_samples = 10
+    h_samples = np.array(
+        [2.2, 6.1, 7.1, 2.9, 3.5, 0.3, 2.9, 2.1, 6.1, 999.5], dtype="float32"
+    )
+    d_samples = cp.asarray(h_samples)
+    num_levels = 7
+    d_histogram = cp.empty(num_levels - 1, dtype="int32")
+    h_num_output_levels = np.array([num_levels], dtype=np.int32)
+    h_lower_level = np.array([0.0], dtype=np.float32)
+    h_upper_level = np.array([12.0], dtype=np.float32)
+
+    builder = make_histogram_even(
+        d_samples=d_samples,
+        d_histogram=d_histogram,
+        h_num_output_levels=h_num_output_levels,
+        h_lower_level=h_lower_level,
+        h_upper_level=h_upper_level,
+        num_samples=num_samples,
+    )
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = deserialize(blob)
+    _run(
+        loaded,
+        d_samples=d_samples,
+        d_histogram=d_histogram,
+        h_num_output_levels=h_num_output_levels,
+        h_lower_level=h_lower_level,
+        h_upper_level=h_upper_level,
+        num_samples=num_samples,
+    )
+
+    expected, _ = np.histogram(
+        h_samples,
+        bins=num_levels - 1,
+        range=(float(h_lower_level[0]), float(h_upper_level[0])),
+    )
+    np.testing.assert_array_equal(d_histogram.get(), expected.astype(np.int32))
