+++ python/cuda_cccl/tests/compute/test_merge_sort.py
@@ -13,8 +13,12 @@ import cuda.compute
 from cuda.compute import (
     CacheModifiedInputIterator,
     OpKind,
+    deserialize,
     gpu_struct,
+    make_merge_sort,
+    serialize,
 )
+from cuda.compute._utils.temp_storage_buffer import TempStorageBuffer
 
 DTYPE_LIST = [
     np.uint8,
@@ -398,3 +402,96 @@ def test_merge_sort_with_values_well_known():
     expected_values = np.array([10, 20, 30, 40])
     np.testing.assert_equal(d_out_keys.get(), expected_keys)
     np.testing.assert_equal(d_out_values.get(), expected_values)
+
+
+def _run(sorter, *, d_in_keys, d_in_values, d_out_keys, d_out_values, num_items, op):
+    bytes_needed = sorter(
+        temp_storage=None,
+        d_in_keys=d_in_keys,
+        d_in_values=d_in_values,
+        d_out_keys=d_out_keys,
+        d_out_values=d_out_values,
+        num_items=num_items,
+        op=op,
+    )
+    tmp = TempStorageBuffer(bytes_needed, None)
+    sorter(
+        temp_storage=tmp,
+        d_in_keys=d_in_keys,
+        d_in_values=d_in_values,
+        d_out_keys=d_out_keys,
+        d_out_values=d_out_values,
+        num_items=num_items,
+        op=op,
+    )
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_merge_sort_keys_values():
+    h_in_keys = np.array([-5, 0, 2, -3, 2, 4, 0, -1, 2, 8], dtype="int32")
+    h_in_values = np.array(
+        [-3.2, 2.2, 1.9, 4.0, -3.9, 2.7, 0, 8.3 - 1, 2.9, 5.4], dtype="float32"
+    )
+    d_in_keys = cp.asarray(h_in_keys)
+    d_in_values = cp.asarray(h_in_values)
+    d_out_keys = cp.empty_like(d_in_keys)
+    d_out_values = cp.empty_like(d_in_values)
+
+    builder = make_merge_sort(
+        d_in_keys=d_in_keys,
+        d_in_values=d_in_values,
+        d_out_keys=d_out_keys,
+        d_out_values=d_out_values,
+        op=OpKind.LESS,
+    )
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = deserialize(blob)
+    _run(
+        loaded,
+        d_in_keys=d_in_keys,
+        d_in_values=d_in_values,
+        d_out_keys=d_out_keys,
+        d_out_values=d_out_values,
+        num_items=d_in_keys.size,
+        op=OpKind.LESS,
+    )
+
+    # kind="stable" works on all supported NumPy versions; the stable= keyword
+    # was only added in NumPy 2.0 and cuda-cccl pins no numpy floor.
+    argsort = np.argsort(h_in_keys, kind="stable")
+    np.testing.assert_array_equal(d_out_keys.get(), h_in_keys[argsort])
+    np.testing.assert_array_equal(d_out_values.get(), h_in_values[argsort])
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_merge_sort_keys_only():
+    # Keys-only: d_in_values / d_out_values are None, which become "none"
+    # iterators — the plain ITER schema members round-trip them fine.
+    h_in_keys = np.array([5, 2, 8, 1, 9, 3, 7, 0, 6, 4], dtype="int32")
+    d_in_keys = cp.asarray(h_in_keys)
+    d_out_keys = cp.empty_like(d_in_keys)
+
+    builder = make_merge_sort(
+        d_in_keys=d_in_keys,
+        d_in_values=None,
+        d_out_keys=d_out_keys,
+        d_out_values=None,
+        op=OpKind.LESS,
+    )
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = deserialize(blob)
+    _run(
+        loaded,
+        d_in_keys=d_in_keys,
+        d_in_values=None,
+        d_out_keys=d_out_keys,
+        d_out_values=None,
+        num_items=d_in_keys.size,
+        op=OpKind.LESS,
+    )
+
+    np.testing.assert_array_equal(d_out_keys.get(), np.sort(h_in_keys))
