+++ python/cuda_cccl/tests/compute/test_scan.py
@@ -14,8 +14,15 @@ from cuda.compute import (
     OpKind,
     ReverseIterator,
     TransformOutputIterator,
+    clear_all_caches,
+    deserialize,
+    exclusive_scan,
     gpu_struct,
+    make_exclusive_scan,
+    make_inclusive_scan,
+    serialize,
 )
+from cuda.compute._utils.temp_storage_buffer import TempStorageBuffer
 
 
 def scan_host(h_input: np.ndarray, op, h_init, force_inclusive):
@@ -509,3 +516,117 @@ def test_scan_bool_maximum(force_inclusive):
         expected = np.array([False, False, True, True], dtype=np.bool_)
 
     np.testing.assert_array_equal(d_output.get(), expected)
+
+
+def _run(scanner, *, d_in, d_out, op, init_value, num_items):
+    bytes_needed = scanner(
+        temp_storage=None,
+        d_in=d_in,
+        d_out=d_out,
+        op=op,
+        init_value=init_value,
+        num_items=num_items,
+    )
+    tmp = TempStorageBuffer(bytes_needed, None)
+    scanner(
+        temp_storage=tmp,
+        d_in=d_in,
+        d_out=d_out,
+        op=op,
+        init_value=init_value,
+        num_items=num_items,
+    )
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_exclusive_scan_round_trip():
+    d_in = cp.arange(1, 33, dtype=cp.int32)
+    d_out = cp.empty_like(d_in)
+    init_value = np.array([0], dtype=np.int32)
+
+    builder = make_exclusive_scan(
+        d_in=d_in, d_out=d_out, op=OpKind.PLUS, init_value=init_value
+    )
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = deserialize(blob)
+    _run(
+        loaded,
+        d_in=d_in,
+        d_out=d_out,
+        op=OpKind.PLUS,
+        init_value=init_value,
+        num_items=d_in.size,
+    )
+
+    expected = np.zeros_like(d_in.get())
+    np.cumsum(d_in.get()[:-1], out=expected[1:])
+    np.testing.assert_array_equal(d_out.get(), expected)
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_inclusive_scan_round_trip():
+    d_in = cp.arange(1, 33, dtype=cp.int32)
+    d_out = cp.empty_like(d_in)
+    init_value = np.array([0], dtype=np.int32)
+
+    builder = make_inclusive_scan(
+        d_in=d_in, d_out=d_out, op=OpKind.PLUS, init_value=init_value
+    )
+    blob = serialize(builder)
+
+    loaded = deserialize(blob)
+    _run(
+        loaded,
+        d_in=d_in,
+        d_out=d_out,
+        op=OpKind.PLUS,
+        init_value=init_value,
+        num_items=d_in.size,
+    )
+
+    np.testing.assert_array_equal(d_out.get(), np.cumsum(d_in.get()))
+
+
+@pytest.mark.serialization
+def test_deserialize_after_jit_matches_jit_result():
+    """Serialize a JITed scan, deserialize, and confirm output matches a fresh JIT."""
+    d_in = cp.array([-5, 0, 2, -3, 2, 4, 0, -1, 2, 8], dtype=cp.int32)
+    d_out_jit = cp.empty_like(d_in)
+    d_out_serialization = cp.empty_like(d_in)
+    init_value = np.array([1], dtype=np.int32)
+
+    def max_op(a, b):
+        return a if a > b else b
+
+    # Build + serialize (this JITs), then clear every in-process cache so the serialization
+    # leg below runs cold: the deserialized scan must stand on its own and cannot
+    # free-ride on a callable warmed by the build or by the JIT reference.
+    blob = serialize(
+        make_exclusive_scan(
+            d_in=d_in, d_out=d_out_serialization, op=max_op, init_value=init_value
+        )
+    )
+    clear_all_caches()
+
+    loaded = deserialize(blob)
+    _run(
+        loaded,
+        d_in=d_in,
+        d_out=d_out_serialization,
+        op=max_op,
+        init_value=init_value,
+        num_items=d_in.size,
+    )
+
+    # Compute the JIT reference only after the serialization path has already run.
+    exclusive_scan(
+        d_in=d_in,
+        d_out=d_out_jit,
+        op=max_op,
+        init_value=init_value,
+        num_items=d_in.size,
+    )
+
+    np.testing.assert_array_equal(d_out_serialization.get(), d_out_jit.get())
