+++ python/cuda_cccl/tests/compute/test_select.py
@@ -7,7 +7,15 @@ import numpy as np
 import pytest
 
 import cuda.compute
-from cuda.compute import CacheModifiedInputIterator, ZipIterator, gpu_struct
+from cuda.compute import (
+    CacheModifiedInputIterator,
+    ZipIterator,
+    deserialize,
+    gpu_struct,
+    make_select,
+    serialize,
+)
+from cuda.compute._utils.temp_storage_buffer import TempStorageBuffer
 
 DTYPE_LIST = [
     np.uint8,
@@ -757,3 +765,49 @@ def test_stateful_caching_same_dtype_different_values():
     # Verify correct results (not cache collision)
     assert count_30 == 69  # Values 31-99
     assert count_70 == 29  # Values 71-99
+
+
+def _even(x):
+    return x % 2 == 0
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_select_round_trip():
+    n = 1024
+    h_in = np.arange(n, dtype=np.int32)
+    d_in = cp.asarray(h_in)
+    d_out = cp.empty_like(d_in)
+    d_num_selected = cp.empty(2, dtype=np.uint64)
+
+    builder = make_select(
+        d_in=d_in, d_out=d_out, d_num_selected_out=d_num_selected, cond=_even
+    )
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = deserialize(blob)
+
+    def _run():
+        nbytes = loaded(
+            temp_storage=None,
+            d_in=d_in,
+            d_out=d_out,
+            d_num_selected_out=d_num_selected,
+            cond=_even,
+            num_items=n,
+        )
+        loaded(
+            temp_storage=TempStorageBuffer(nbytes, None),
+            d_in=d_in,
+            d_out=d_out,
+            d_num_selected_out=d_num_selected,
+            cond=_even,
+            num_items=n,
+        )
+
+    _run()
+
+    k = int(d_num_selected[0].get())
+    expected = h_in[h_in % 2 == 0]
+    assert k == expected.size
+    np.testing.assert_array_equal(d_out.get()[:k], expected)
