+++ python/cuda_cccl/tests/compute/test_transform.py
@@ -10,7 +10,11 @@ import cuda.compute
 from cuda.compute import (
     CountingIterator,
     OpKind,
+    deserialize,
     gpu_struct,
+    make_binary_transform,
+    make_unary_transform,
+    serialize,
 )
 
 
@@ -704,3 +708,43 @@ def test_transform_caching_with_global_np_ufunc():
     cp.testing.assert_allclose(d_out, cp.cos(d_in))
 
     d_in = cp.asarray([1.0, 2.0, 3.0])
+
+
+def _add_one(a):
+    return a + 1
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_unary_transform_round_trip():
+    h_in = np.array([1, 2, 3, 4, 5], dtype=np.int32)
+    d_in = cp.asarray(h_in)
+    d_out = cp.empty_like(d_in)
+
+    builder = make_unary_transform(d_in=d_in, d_out=d_out, op=_add_one)
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = deserialize(blob)
+    loaded(d_in=d_in, d_out=d_out, op=_add_one, num_items=d_in.size)
+
+    np.testing.assert_array_equal(d_out.get(), h_in + 1)
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_binary_transform_round_trip():
+    h_in1 = np.array([1, 2, 3, 4], dtype=np.int32)
+    h_in2 = np.array([10, 20, 30, 40], dtype=np.int32)
+    d_in1 = cp.asarray(h_in1)
+    d_in2 = cp.asarray(h_in2)
+    d_out = cp.empty_like(d_in1)
+
+    builder = make_binary_transform(
+        d_in1=d_in1, d_in2=d_in2, d_out=d_out, op=OpKind.PLUS
+    )
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = deserialize(blob)
+    loaded(d_in1=d_in1, d_in2=d_in2, d_out=d_out, op=OpKind.PLUS, num_items=d_in1.size)
+
+    np.testing.assert_array_equal(d_out.get(), h_in1 + h_in2)
