+++ python/cuda_cccl/tests/compute/test_unique_by_key.py
@@ -13,8 +13,12 @@ from cuda.compute import (
     CacheModifiedInputIterator,
     DiscardIterator,
     OpKind,
+    deserialize,
     gpu_struct,
+    make_unique_by_key,
+    serialize,
 )
+from cuda.compute._utils.temp_storage_buffer import TempStorageBuffer
 
 DTYPE_LIST = [
     np.uint8,
@@ -457,3 +461,78 @@ def test_unique_by_key_well_known_equal_to(monkeypatch):
 
     np.testing.assert_equal(d_out_keys.get()[:3], expected_keys)
     np.testing.assert_equal(d_out_values.get()[:3], expected_values)
+
+
+def _run(
+    uniquer,
+    *,
+    d_in_keys,
+    d_in_values,
+    d_out_keys,
+    d_out_values,
+    d_out_num_selected,
+    num_items,
+    op,
+):
+    bytes_needed = uniquer(
+        temp_storage=None,
+        d_in_keys=d_in_keys,
+        d_in_items=d_in_values,
+        d_out_keys=d_out_keys,
+        d_out_items=d_out_values,
+        d_out_num_selected=d_out_num_selected,
+        op=op,
+        num_items=num_items,
+    )
+    tmp = TempStorageBuffer(bytes_needed, None)
+    uniquer(
+        temp_storage=tmp,
+        d_in_keys=d_in_keys,
+        d_in_items=d_in_values,
+        d_out_keys=d_out_keys,
+        d_out_items=d_out_values,
+        d_out_num_selected=d_out_num_selected,
+        op=op,
+        num_items=num_items,
+    )
+
+
+@pytest.mark.serialization
+def test_serialize_deserialize_unique_by_key_round_trip():
+    h_in_keys = np.array([0, 2, 2, 9, 5, 5, 5, 8], dtype="int32")
+    h_in_values = np.array([1, 2, 3, 4, 5, 6, 7, 8], dtype="float32")
+
+    d_in_keys = cp.asarray(h_in_keys)
+    d_in_values = cp.asarray(h_in_values)
+    d_out_keys = cp.empty_like(d_in_keys)
+    d_out_values = cp.empty_like(d_in_values)
+    d_out_num_selected = cp.empty(1, np.int32)
+
+    builder = make_unique_by_key(
+        d_in_keys=d_in_keys,
+        d_in_items=d_in_values,
+        d_out_keys=d_out_keys,
+        d_out_items=d_out_values,
+        d_out_num_selected=d_out_num_selected,
+        op=OpKind.EQUAL_TO,
+    )
+    blob = serialize(builder)
+    assert len(blob) > 0
+
+    loaded = deserialize(blob)
+    _run(
+        loaded,
+        d_in_keys=d_in_keys,
+        d_in_values=d_in_values,
+        d_out_keys=d_out_keys,
+        d_out_values=d_out_values,
+        d_out_num_selected=d_out_num_selected,
+        num_items=d_in_keys.size,
+        op=OpKind.EQUAL_TO,
+    )
+
+    n = int(d_out_num_selected.get()[0])
+    np.testing.assert_array_equal(d_out_keys.get()[:n], np.array([0, 2, 9, 5, 8]))
+    np.testing.assert_array_equal(
+        d_out_values.get()[:n], np.array([1, 2, 4, 5, 8], dtype=np.float32)
+    )
